api/requirements.txt
autopep8==1.5.6
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
protobuf==3.15.6
pycodestyle==2.7.0
python-dotenv==0.15.0
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
